+++ baseapp/options.go
@@ -95,21 +95,23 @@ func (app *BaseApp) SetParamStore(ps ParamStore) {
 	app.paramStore = ps
 }
 
-// SetAppVersion sets the application's version string.
-func (app *BaseApp) SetAppVersion(v string) {
+// SetVersion sets the application's version string.
+func (app *BaseApp) SetVersion(v string) {
 	if app.sealed {
-		panic("SetAppVersion() on sealed BaseApp")
+		panic("SetVersion() on sealed BaseApp")
 	}
-
-	app.appVersion = v
+	app.version = v
 }
 
-type ProtocolManager interface {
+// ProtocolVersionManager exposes functionality to set
+// BaseApp's protocol version.
+type ProtocolVersionManager interface {
 	SetProtocolVersion(uint64)
 }
 
+// SetProtocolVersion sets the application's protocol version
 func (app *BaseApp) SetProtocolVersion(v uint64) {
-	app.protocolVersion = v
+	app.appVersion = v
 }
 
 func (app *BaseApp) SetDB(db dbm.DB) {